feat(0.4.0): release 0.4.0
lint and conformance working in build suite, additional issues resolved

+++ contracts/Certificate.sol
@@ -9,7 +9,6 @@ import "./interfaces/FileInterface.sol";
 import "./interfaces/EventInterface.sol";
 import "./libraries/Utils.sol";
 
-
 contract Certificate is CertificateInterface {
     address public certifiedEmail;
     address public owner;
@@ -19,37 +18,25 @@ contract Certificate is CertificateInterface {
 
     string[] public eventsId;
 
-    uint public createdAt;
+    uint256 public createdAt;
 
     FileInterface public file;
 
     mapping(string => EventInterface) private events;
 
-    constructor(
-        string memory certificateId,
-        address certifiedEmailDeployer
-    ) public {
+    constructor(string memory certificateId, address certifiedEmailDeployer) public {
         id = certificateId;
         deployer = certifiedEmailDeployer;
         certifiedEmail = msg.sender;
     }
 
     modifier certifiedEmailModifier() {
-        require(
-            msg.sender == certifiedEmail,
-            "Only CertifiedEmail account can perform this action"
-        );
+        require(msg.sender == certifiedEmail, "Only CertifiedEmail account can perform this action");
 
         _;
     }
 
-    function init(
-        address certificateOwner,
-        uint certificateCreatedAt
-    )
-        public
-        certifiedEmailModifier
-    {
+    function init(address certificateOwner, uint256 certificateCreatedAt) public certifiedEmailModifier {
         owner = certificateOwner;
         createdAt = certificateCreatedAt;
     }
@@ -58,85 +45,52 @@ contract Certificate is CertificateInterface {
         string memory fileHash,
         string memory fileId,
         string memory fileName,
-        uint fileCreatedAt,
-        uint fileSize
-    )
-        public
-        certifiedEmailModifier
-    {
-        (bool success, bytes memory returnData) = deployer.delegatecall(
-            abi.encodeWithSignature(
-                "deployFile(string)",
-                fileId
-            )
-        );
+        uint256 fileCreatedAt,
+        uint256 fileSize
+    ) public certifiedEmailModifier {
+        (bool success, bytes memory returnData) =
+            deployer.delegatecall(abi.encodeWithSignature("deployFile(string)", fileId));
 
-        require(
-            success,
-            "Error while deploying file from certificate"
-        );
+        require(success, "Error while deploying file from certificate");
 
         file = FileInterface(Utils._bytesToAddress(returnData));
 
-        file.init(
-                fileName,
-                fileHash,
-                fileCreatedAt,
-                fileSize
-        );
+        file.init(fileName, fileHash, fileCreatedAt, fileSize);
     }
 
     function createEvent(
         string memory eventId,
         string memory eventType,
         string memory eventUserAgent,
-        uint eventCreatedAt
-    )
-        public
-        certifiedEmailModifier
-    {
-        require(
-            address(events[eventId]) == address(0),
-            "The current event already exists"
-        );
+        uint256 eventCreatedAt
+    ) public certifiedEmailModifier {
+        require(address(events[eventId]) == address(0), "The current event already exists");
 
-        (bool success, bytes memory returnData) = deployer.delegatecall(
-            abi.encodeWithSignature(
-                "deployEvent(string,string,string,uint256)",
-                eventId,
-                eventType,
-                eventUserAgent,
-                eventCreatedAt
-            )
-        );
+        (bool success, bytes memory returnData) =
+            deployer.delegatecall(
+                abi.encodeWithSignature(
+                    "deployEvent(string,string,string,uint256)",
+                    eventId,
+                    eventType,
+                    eventUserAgent,
+                    eventCreatedAt
+                )
+            );
 
-        require(
-            success,
-            "Error while deploying event from certificate"
-        );
+        require(success, "Error while deploying event from certificate");
 
         events[eventId] = EventInterface(Utils._bytesToAddress(returnData));
 
         eventsId.push(eventId);
     }
 
-    function getEvent(
-        string memory eventId
-    )
-        public
-        view
-        returns (address)
-    {
+    function getEvent(string memory eventId) public view returns (address) {
         if (address(events[eventId]) == address(0)) return address(0);
 
         return address(events[eventId]);
     }
 
-    function getEventsSize()
-        public
-        view
-        returns (uint)
-    {
+    function getEventsSize() public view returns (uint256) {
         return eventsId.length;
     }
 }
