+++ contracts/CertifiedEmailDeployer.sol
@@ -8,27 +8,14 @@ import "./File.sol";
 import "./Certificate.sol";
 import "./Event.sol";
 
-
 contract CertifiedEmailDeployer {
-
     constructor() public {}
 
-    function deployFile(
-        string memory fileId
-    )
-        public
-        returns (address)
-    {
+    function deployFile(string memory fileId) public returns (address) {
         return address(new File(fileId));
     }
 
-    function deployCertificate(
-        string memory certificateId,
-        address deployer
-    )
-        public
-        returns (address)
-    {
+    function deployCertificate(string memory certificateId, address deployer) public returns (address) {
         return address(new Certificate(certificateId, deployer));
     }
 
@@ -36,18 +23,8 @@ contract CertifiedEmailDeployer {
         string memory eventId,
         string memory eventType,
         string memory eventUserAgent,
-        uint eventCreatedAt
-    )
-        public
-        returns (address)
-    {
-        return address(
-            new Event(
-                eventId,
-                eventType,
-                eventUserAgent,
-                eventCreatedAt
-            )
-        );
+        uint256 eventCreatedAt
+    ) public returns (address) {
+        return address(new Event(eventId, eventType, eventUserAgent, eventCreatedAt));
     }
 }
