+++ contracts/Migrations.sol
@@ -1,35 +1,22 @@
 pragma solidity <0.6.0;
 
-
 contract Migrations {
     address public owner;
-    uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
     constructor() public {
         owner = msg.sender;
     }
 
     modifier restricted() {
-        if (msg.sender == owner)
-
-            _;
+        if (msg.sender == owner) _;
     }
 
-    function setCompleted(
-        uint completed
-    )
-        public
-        restricted
-    {
+    function setCompleted(uint256 completed) public restricted {
         lastCompletedMigration = completed;
     }
 
-    function upgrade(
-        address new_address
-    )
-        public
-        restricted
-    {
+    function upgrade(address new_address) public restricted {
         Migrations upgraded = Migrations(new_address);
         upgraded.setCompleted(lastCompletedMigration);
     }
