+++ contracts/interfaces/CertifiedFileCheckerInterface.sol
@@ -5,18 +5,15 @@ import "./NotifierInterface.sol";
 contract CertifiedFileCheckerInterface is NotifierInterface {
     address public as2network;
 
-    function getFile(
-        string memory fileHash,
-        uint index
-    )
+    function getFile(string memory fileHash, uint256 index)
         public
         view
-        returns(
+        returns (
             string memory id,
             string memory hash,
-            uint size,
-            uint createdAt,
-            address  owner,
+            uint256 size,
+            uint256 createdAt,
+            address owner,
             address contract_address,
             bool more
         );
