+++ contracts/interfaces/DocumentInterface.sol
@@ -3,7 +3,6 @@ pragma solidity <0.6.0;
 import "./FileInterface.sol";
 import "./EventInterface.sol";
 
-
 contract DocumentInterface {
     address public signature;
     address public signer;
@@ -17,8 +16,8 @@ contract DocumentInterface {
 
     string[] public eventsId;
 
-    uint public signedAt;
-    uint public createdAt;
+    uint256 public signedAt;
+    uint256 public createdAt;
 
     bool public signed;
     bool public canceled;
@@ -28,75 +27,42 @@ contract DocumentInterface {
 
     mapping(string => EventInterface) private events;
 
-    function init(
-        string memory initType,
-        uint documentCreatedAt
-    )
-        public;
+    function init(string memory initType, uint256 documentCreatedAt) public;
 
-    function setSignatureOwner(
-        address signatureOwnerAdr
-    )
-        public;
+    function setSignatureOwner(address signatureOwnerAdr) public;
 
-    function setOwner(
-        address signerAddress
-    )
-        public;
+    function setOwner(address signerAddress) public;
 
-    function sign(
-        uint documentSignedAt
-    )
-        public;
+    function sign(uint256 documentSignedAt) public;
 
     function notifyEntityEvent(
         string memory notifiersKey,
         string memory createdEvent,
         address adrToNotify
-    )
-        internal;
+    ) internal;
 
-    function decline(
-        string memory documentDeclineReason
-    )
-        public;
+    function decline(string memory documentDeclineReason) public;
 
-    function cancel(
-        string memory documentCancelReason
-    )
-        public;
+    function cancel(string memory documentCancelReason) public;
 
     function createFile(
         string memory fileId,
         string memory fileName,
         string memory fileHash,
-        uint fileCreatedAt,
-        uint fileSize
-    )
-        public;
+        uint256 fileCreatedAt,
+        uint256 fileSize
+    ) public;
 
-    function setFileHash(
-        string memory fileHash
-    )
-        public;
+    function setFileHash(string memory fileHash) public;
 
     function createEvent(
         string memory eventId,
         string memory eventType,
         string memory eventUserAgent,
-        uint eventCreatedAt
-    )
-        public;
+        uint256 eventCreatedAt
+    ) public;
 
-    function getEvent(
-        string memory eventId
-    )
-        public
-        view
-        returns (address);
+    function getEvent(string memory eventId) public view returns (address);
 
-    function getEventsSize()
-        public
-        view
-        returns (uint);
+    function getEventsSize() public view returns (uint256);
 }
