+++ contracts/interfaces/EventInterface.sol
@@ -1,9 +1,8 @@
 pragma solidity <0.6.0;
 
-
 contract EventInterface {
     string public id;
     string public eventType;
     string public userAgent;
-    uint public createdAt;
+    uint256 public createdAt;
 }
