+++ contracts/libraries/Utils.sol
@@ -1,30 +1,15 @@
 pragma solidity <0.6.0;
 
-
 library Utils {
-    function _bytesToAddress(
-        bytes memory rawData
-    )
-        internal
-        pure
-        returns (address)
-    {
-        uint bytesToUint;
+    function _bytesToAddress(bytes memory rawData) internal pure returns (address) {
+        uint256 bytesToUint;
         assembly {
-            bytesToUint := mload(add(rawData,0x20))
+            bytesToUint := mload(add(rawData, 0x20))
         }
         return address(uint160(uint256(bytesToUint)));
     }
 
-    function keccak (
-        string memory key
-    )
-        internal
-        pure
-        returns (bytes32)
-    {
-        return keccak256(
-            abi.encode(key)
-        );
+    function keccak(string memory key) internal pure returns (bytes32) {
+        return keccak256(abi.encode(key));
     }
 }
